refactor: decompose OllamaBackend
