+++ Sources/ManifoldCloud/OllamaBackend.swift
@@ -216,9 +216,13 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
         let planned = plan.effectiveContextSize
         effectiveNumCtx = planned > Self.defaultNumCtxFloor ? planned : Self.defaultNumCtxFloor
 
-        let probed: ShowProbe
+        let probed: OllamaShowProbe
         do {
-            probed = try await probeShow()
+            probed = try await OllamaModelProbe.probeShow(
+                baseURL: configuredBaseURL,
+                modelName: modelName,
+                urlSession: urlSession
+            )
         } catch {
             Log.network.info("OllamaBackend /api/show probe threw \(error.localizedDescription, privacy: .public) — treating \(self.modelName, privacy: .public) as non-thinking with conservative manifest")
             probed = .empty
@@ -247,114 +251,6 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
 
         setIsModelLoaded(true)
         Log.inference.info("OllamaBackend configured for \(self.modelName, privacy: .public) at \(self.baseURL?.host() ?? "unknown", privacy: .public) thinking=\(self.isThinkingModel, privacy: .public) num_ctx=\(self.effectiveNumCtx, privacy: .public) ctxWindow=\(resolvedContextWindow, privacy: .public)")
-    }
-
-    /// Decoded subset of Ollama's `/api/show` response we actually consume.
-    private struct ShowProbe {
-        let thinking: Bool
-        let thinkingMarkers: ThinkingMarkers?
-        let contextLength: Int?
-
-        static let empty = ShowProbe(thinking: false, thinkingMarkers: nil, contextLength: nil)
-    }
-
-    /// Calls Ollama's `/api/show` endpoint and extracts thinking capability,
-    /// auto-detected thinking markers, and the model's true context window.
-    ///
-    /// Thinking detection prefers `capabilities: ["thinking", ...]` (surfaced
-    /// by modern Ollama releases) and falls back to scanning the Jinja
-    /// `template` field for `<think>` / `{{ if .Thinking }}` markers that
-    /// reasoning models ship by convention. ``ThinkingMarkers`` are extracted
-    /// from the same template so the wire-format inline-fallback path
-    /// (`OllamaBackend.parseResponseStream`) can route reasoning content
-    /// through ``ThinkingParser`` rather than hardcoding `<think>` even when
-    /// the loaded model uses a different marker family.
-    ///
-    /// Context length is read from `model_info.context_length`. On HTTP
-    /// failures or shape mismatches the probe returns ``ShowProbe/empty``;
-    /// only true network exceptions throw, so callers can wrap the call in
-    /// `try?` for best-effort behaviour.
-    private func probeShow() async throws -> ShowProbe {
-        guard let baseURL else { return .empty }
-        let showURL = baseURL.appendingPathComponent("api/show")
-
-        var request = URLRequest(url: showURL)
-        request.httpMethod = "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.httpBody = try JSONSerialization.data(withJSONObject: ["model": modelName])
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await urlSession.data(for: request)
-        } catch {
-            Log.network.info("OllamaBackend /api/show probe failed (\(error.localizedDescription, privacy: .public)) — treating \(self.modelName, privacy: .public) as non-thinking")
-            return .empty
-        }
-
-        if let http = response as? HTTPURLResponse, !(200...299).contains(http.statusCode) {
-            Log.network.info("OllamaBackend /api/show returned HTTP \(http.statusCode, privacy: .public) for \(self.modelName, privacy: .public) — treating as non-thinking")
-            return .empty
-        }
-
-        let json: [String: Any]
-        do {
-            guard let parsed = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                Log.network.info("OllamaBackend /api/show returned non-object JSON for \(self.modelName, privacy: .public) — treating as non-thinking")
-                return .empty
-            }
-            json = parsed
-        } catch {
-            Log.network.info("OllamaBackend /api/show returned non-JSON for \(self.modelName, privacy: .public) — treating as non-thinking")
-            return .empty
-        }
-
-        // Thinking capability — prefer the structured capabilities list.
-        var thinking = false
-        if let caps = json["capabilities"] as? [String],
-           caps.contains(where: { $0.lowercased() == "thinking" }) {
-            thinking = true
-        }
-
-        // Thinking markers — auto-detect from the Jinja template via the
-        // shared ``PromptTemplateDetector`` rules. Then opportunistically
-        // back-fill the thinking flag when the template carries a
-        // recognised marker pair.
-        var detectedMarkers: ThinkingMarkers?
-        if let template = json["template"] as? String {
-            detectedMarkers = ThinkingMarkers.fromChatTemplate(template)
-            if !thinking {
-                let legacyMarkers = ["<think>", "</think>", "{{ if .Thinking }}", "{{if .Thinking}}"]
-                if legacyMarkers.contains(where: { template.contains($0) }) {
-                    thinking = true
-                }
-            }
-        }
-
-        // model_info.context_length — the authoritative wire value. Missing
-        // on snapshots compiled from older Modelfiles; fall back to the
-        // caller's plan-derived window in that case.
-        var contextLength: Int?
-        if let modelInfo = json["model_info"] as? [String: Any] {
-            // Ollama reports context length under either the canonical
-            // `context_length` key or the architecture-prefixed
-            // `<arch>.context_length` (e.g. `llama.context_length`).
-            for (key, value) in modelInfo where key.hasSuffix("context_length") {
-                if let int = value as? Int, int > 0 {
-                    contextLength = int
-                    break
-                }
-                if let int64 = value as? Int64, int64 > 0 {
-                    contextLength = Int(int64)
-                    break
-                }
-            }
-        }
-
-        return ShowProbe(
-            thinking: thinking,
-            thinkingMarkers: detectedMarkers,
-            contextLength: contextLength
-        )
     }
 
     // MARK: - Request Building
@@ -393,7 +289,7 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
             return snapshot
         }
         if let toolHistory = snapshotToolHistory {
-            messages.append(contentsOf: toolHistory.map(Self.encodeToolAwareEntry))
+            messages.append(contentsOf: toolHistory.map(OllamaMessageEncoder.encodeToolAwareEntry))
         } else if let history = conversationHistory {
             messages.append(contentsOf: history.map { ["role": $0.role, "content": $0.content] })
         } else {
@@ -480,20 +376,8 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
         // are passed through as literal strings; `.tool(name:)` produces the
         // function-selection object Ollama expects for forced selection.
         if !config.tools.isEmpty {
-            body["tools"] = config.tools.map(Self.encodeToolDefinition)
-            switch config.toolChoice {
-            case .auto:
-                break
-            case .none:
-                body["tool_choice"] = "none"
-            case .required:
-                body["tool_choice"] = "required"
-            case .tool(let name):
-                body["tool_choice"] = [
-                    "type": "function",
-                    "function": ["name": name],
-                ]
-            }
+            body["tools"] = config.tools.map(OllamaMessageEncoder.encodeToolDefinition)
+            OllamaMessageEncoder.applyToolChoice(config.toolChoice, into: &body)
         }
 
         var request = URLRequest(url: chatURL)
@@ -515,360 +399,19 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
     // timeout/2 partially addresses this by showing .stalled.
 
     /// Parses Ollama's NDJSON response format instead of SSE.
-    ///
-    /// Applies the same ``SSEStreamLimits`` caps as the SSE parser so a
-    /// hostile Ollama-compatible server cannot exhaust memory with oversized
-    /// lines, total volume, or an event flood.
-    ///
-    /// Reasoning models (qwen3, qwen3.5:4b, deepseek-r1) surface chain-of-thought
-    /// tokens in a separate `thinking` field — `message.thinking` on the
-    /// `/api/chat` endpoint and top-level `thinking` on `/api/generate`. We
-    /// emit ``GenerationEvent/thinkingToken(_:)`` while a line carries
-    /// non-empty thinking, and ``GenerationEvent/thinkingComplete`` exactly
-    /// once — either on the transition from "thinking was non-empty" to
-    /// "thinking is now empty", or on `"done":true` when a thinking
-    /// accumulator is still open. ``GenerationConfig/maxThinkingTokens``
-    /// caps reasoning emission; once exceeded subsequent thinking content is
-    /// dropped and only visible ``GenerationEvent/token(_:)`` events continue.
-    ///
-    /// Fallback for models that leak reasoning into content: some Ollama
-    /// models (e.g. Qwen3 tags that don't populate `message.thinking` on this
-    /// server) emit `<think>…</think>` blocks inline in `content` instead.
-    /// When we never see a populated `thinking` field on the stream and a
-    /// content chunk contains the opening marker, content is routed through
-    /// ``ThinkingParser`` so callers still receive
-    /// ``GenerationEvent/thinkingToken(_:)`` /
-    /// ``GenerationEvent/thinkingComplete`` events rather than the raw tags.
-    /// The ``GenerationConfig/maxThinkingTokens`` cap still applies; visible
-    /// content emerges from the parser as ``GenerationEvent/token(_:)``.
-    ///
-    /// Limitation: engagement is per-chunk, so an opening marker split across
-    /// two NDJSON content chunks (e.g. `<th` + `ink>`) would miss detection
-    /// and yield raw tag fragments as visible tokens. In practice Ollama
-    /// emits `message.content` in coarse line-sized chunks, so the opening
-    /// tag lands in a single chunk. Once engaged, the parser's own buffering
-    /// correctly reassembles a closing tag split across chunks.
     public override func parseResponseStream(
         bytes: URLSession.AsyncBytes,
         config: GenerationConfig,
         continuation: AsyncThrowingStream<GenerationEvent, Error>.Continuation
     ) async throws {
-        let limits = effectiveSSEStreamLimits
-        var lineBuffer = Data()
-        var totalBytes = 0
-        var rateWindowStart = ContinuousClock.now
-        var rateWindowCount = 0
-
-        // Tracks whether we've emitted any thinking content on this stream,
-        // so we know when to fire the single .thinkingComplete event.
-        var thinkingOpen = false
-        var thinkingTokenCount = 0
-        let thinkingLimit = config.maxThinkingTokens
-        // Fallback per-line counter for servers that don't emit `eval_count`
-        // until the done-line. When a line carries `eval_count`, we prefer it
-        // over this counter for an exact cap.
-        var visibleLineCount = 0
-        let visibleLimit = config.maxOutputTokens
-
-        // Inline `<think>` fallback state. Engaged only when the server never
-        // populates `message.thinking` / top-level `thinking` and a content
-        // chunk carries `<think>`. Once engaged it stays engaged for the rest
-        // of the stream so partial tags split across chunks are held back
-        // correctly by the parser's own buffering.
-        var sawThinkingField = false
-        // Order: explicit per-request override > auto-detected markers from
-        // the `/api/show` template scan > Qwen3 default. Hardcoding
-        // `.qwen3` was the bug — non-Qwen reasoning models (`<thinking>`,
-        // `<reasoning>`) silently leaked their markers as visible tokens.
-        let detectedMarkers = withStateLock { _autoDetectedThinkingMarkers }
-        let fallbackMarkers = config.thinkingMarkers ?? detectedMarkers ?? .qwen3
-        var contentParser: ThinkingParser?
-
-        func noteEventYielded() throws {
-            let now = ContinuousClock.now
-            if now - rateWindowStart >= .seconds(1) {
-                rateWindowStart = now
-                rateWindowCount = 1
-                return
-            }
-            rateWindowCount += 1
-            if rateWindowCount > limits.maxEventsPerSecond {
-                throw SSEStreamError.eventRateExceeded(rateWindowCount)
-            }
-        }
-
-        // Yield a single parser-produced event while honouring the per-stream
-        // caps used elsewhere. Returns `false` when the visible-token cap was
-        // hit and the caller should stop producing further output on this
-        // line. Only handles the events `ThinkingParser` actually emits
-        // (`.token`, `.thinkingToken`, `.thinkingComplete`); anything else is
-        // forwarded verbatim.
-        func emit(_ event: GenerationEvent) throws -> Bool {
-            switch event {
-            case .thinkingToken(let text):
-                if let limit = thinkingLimit, thinkingTokenCount >= limit {
-                    return true // Drop silently — cap reached.
-                }
-                try noteEventYielded()
-                continuation.yield(.thinkingToken(text))
-                thinkingOpen = true
-                thinkingTokenCount += 1
-                return true
-            case .thinkingComplete:
-                try noteEventYielded()
-                continuation.yield(.thinkingComplete)
-                thinkingOpen = false
-                return true
-            case .token(let text):
-                if let limit = visibleLimit, visibleLineCount >= limit {
-                    continuation.finish()
-                    return false
-                }
-                try noteEventYielded()
-                continuation.yield(.token(text))
-                visibleLineCount += 1
-                return true
-            default:
-                continuation.yield(event)
-                return true
-            }
-        }
-
-        // Tool calls first: Ollama can emit multiple tool_calls in a single
-        // assistant message. Dispatch them in emission order so the
-        // coordinator's serial dispatch loop sees them in the same order
-        // the model produced them.
-        //
-        // Event-shape contract (PR #783): every tool call surfaces as a
-        // uniform start + single arguments-delta + toolCall triple, even
-        // for whole-call backends like Ollama. That keeps consumers
-        // (orchestrator, UI) on a single code path regardless of whether
-        // the underlying transport streams arguments incrementally.
-        //
-        // TODO(#753): Some Ollama configs (notably `qwen2.5:7b` against
-        // newer servers) reportedly emit incremental tool_call deltas
-        // across multiple NDJSON lines. v1 treats Ollama as whole-call
-        // only; if we observe incremental deltas in the wild, lift
-        // `StreamingToolCallAccumulator` from `OpenAIToolEncoding.swift`
-        // and key by tool-call index here. `streamsToolCallArguments`
-        // would flip to `true` at the same time.
-        // Returns `false` when `Task.isCancelled` is observed at any point
-        // during this helper so `handleLine` can skip the remaining
-        // sub-handlers (`processThinking` / `processContent`) for the same
-        // NDJSON line. Pre-#970 the equivalent in-loop `if Task.isCancelled
-        // { return }` exited `handleLine` directly; after the decomposition
-        // a bare `return` exits only this helper, which would let
-        // post-cancel thinking and content events surface — the regression
-        // #972 fixes by mirroring `processContent`'s `Bool` convention.
-        func processToolCalls(_ parsed: ParsedLine) throws -> Bool {
-            guard let toolCalls = parsed.toolCalls, !toolCalls.isEmpty else {
-                // Even with no tool_calls on this line, a cancel that
-                // arrived between sub-handlers must short-circuit
-                // downstream emission for the same line.
-                return !Task.isCancelled
-            }
-            for call in toolCalls {
-                // Cancellation contract: a consumer that drops the stream
-                // mid-flight must NOT observe `.toolCall` events for
-                // entries the orchestrator never agreed to dispatch. Honour
-                // `Task.isCancelled` at the emit boundary so a single-line
-                // tool_calls payload arriving alongside the cancel doesn't
-                // fire a phantom dispatch.
-                if Task.isCancelled { return false }
-                try noteEventYielded()
-                continuation.yield(.toolCallStart(callId: call.id, name: call.toolName))
-                if !call.arguments.isEmpty {
-                    try noteEventYielded()
-                    continuation.yield(.toolCallArgumentsDelta(
-                        callId: call.id,
-                        textDelta: call.arguments
-                    ))
-                }
-                try noteEventYielded()
-                continuation.yield(.toolCall(call))
-            }
-            // Post-loop cancellation: if the cancel arrived after the last
-            // tool_call was emitted but before this helper returned, still
-            // bail so thinking/content on the same line don't surface.
-            return !Task.isCancelled
-        }
-
-        // Route thinking field (if any) first so downstream consumers see
-        // reasoning before visible content for a given NDJSON record.
-        func processThinking(_ parsed: ParsedLine) throws {
-            if let thinking = parsed.thinking, !thinking.isEmpty {
-                sawThinkingField = true
-                if let limit = thinkingLimit, thinkingTokenCount >= limit {
-                    return // Cap reached — drop this thinking chunk silently.
-                }
-                try noteEventYielded()
-                continuation.yield(.thinkingToken(thinking))
-                thinkingOpen = true
-                // Count each thinking-bearing NDJSON line as one "token"
-                // for cap purposes. Ollama ships whole-blob thinking per
-                // line rather than per-token, so this matches the coarser
-                // grain of the wire format.
-                thinkingTokenCount += 1
-                return
-            }
-            // Transition from thinking → content. Fire .thinkingComplete
-            // exactly once on the first empty-thinking line we see after
-            // any non-empty thinking was emitted. Skipped when the fallback
-            // parser is driving state — the parser closes its own thinking
-            // block via its own `.thinkingComplete`.
-            if thinkingOpen && contentParser == nil {
-                try noteEventYielded()
-                continuation.yield(.thinkingComplete)
-                thinkingOpen = false
-            }
-        }
-
-        // Returns `false` when the stream should stop (visible-token cap hit
-        // or the fallback parser surfaced a stop signal); `true` to keep
-        // processing further fields on the same line.
-        func processContent(_ parsed: ParsedLine) throws -> Bool {
-            guard let content = parsed.content, !content.isEmpty else { return true }
-            // Prefer the server's running `eval_count` when present for an
-            // exact token-count cap; otherwise fall back to the NDJSON line
-            // counter which is an upper bound but may overshoot by one line.
-            if let limit = visibleLimit {
-                let observed = parsed.evalCount ?? visibleLineCount
-                if observed >= limit {
-                    continuation.finish()
-                    return false
-                }
-            }
-            // Engage the fallback `<think>`-in-content parser when the
-            // server has never populated a dedicated thinking field on this
-            // stream and the incoming content carries the opening tag. Once
-            // engaged, every subsequent content chunk flows through the
-            // parser so a tag split across two NDJSON lines is held in the
-            // parser's own buffer.
-            if contentParser == nil,
-               !sawThinkingField,
-               content.contains(fallbackMarkers.open) {
-                contentParser = ThinkingParser(markers: fallbackMarkers)
-            }
-            if var parser = contentParser {
-                for event in parser.process(content) {
-                    if try !emit(event) {
-                        contentParser = parser
-                        return false
-                    }
-                }
-                contentParser = parser
-            } else {
-                try noteEventYielded()
-                continuation.yield(.token(content))
-                visibleLineCount += 1
-            }
-            return true
-        }
-
-        // Returns `false` when the fallback parser surfaced a stop signal
-        // mid-flush (visible-token cap hit while draining held-back bytes).
-        func processDoneFlush() throws -> Bool {
-            // Flush any remaining buffered content from the fallback parser
-            // first — held-back bytes (e.g. an unmatched prefix of `<`)
-            // must be emitted before we decide whether thinking is still
-            // open.
-            if var parser = contentParser {
-                for event in parser.finalize() {
-                    if try !emit(event) {
-                        contentParser = parser
-                        return false
-                    }
-                }
-                contentParser = parser
-            }
-            // Ollama can terminate with `"done":true` while thinking is
-            // still the only content emitted (e.g. reasoning model hits
-            // num_predict mid-think). Flush .thinkingComplete so downstream
-            // consumers don't leave the thinking block open.
-            if thinkingOpen {
-                try noteEventYielded()
-                continuation.yield(.thinkingComplete)
-                thinkingOpen = false
-            }
-            return true
-        }
-
-        // Surface usage from the done-line (`eval_count`,
-        // `prompt_eval_count`). Wires into `handleUsage` (populating
-        // `lastUsage` for `TokenUsageProvider` consumers) and emits a
-        // `.usage` event on the stream, mirroring the SSE path in
-        // `SSECloudBackend.parseResponseStream`.
-        func processUsage(_ parsed: ParsedLine) throws {
-            guard parsed.evalCount != nil || parsed.promptEvalCount != nil else { return }
-            let usage: (promptTokens: Int?, completionTokens: Int?) = (
-                promptTokens: parsed.promptEvalCount,
-                completionTokens: parsed.evalCount
-            )
-            handleUsage(usage)
-            if let prompt = usage.promptTokens,
-               let completion = usage.completionTokens {
-                try noteEventYielded()
-                continuation.yield(.usage(prompt: prompt, completion: completion))
-            }
-        }
-
-        func handleLine(_ line: String) throws {
-            guard let parsed = Self.parseLine(line) else { return }
-            guard try processToolCalls(parsed) else { return }
-            try processThinking(parsed)
-            guard try processContent(parsed) else { return }
-            if parsed.done {
-                guard try processDoneFlush() else { return }
-                try processUsage(parsed)
-            }
-        }
-
-        for try await byte in bytes {
-            if Task.isCancelled { break }
-
-            totalBytes += 1
-            if totalBytes > limits.maxTotalBytes {
-                throw SSEStreamError.streamTooLarge(totalBytes)
-            }
-
-            if byte == UInt8(ascii: "\n") {
-                if !lineBuffer.isEmpty {
-                    if let line = String(data: lineBuffer, encoding: .utf8) {
-                        try handleLine(line)
-                    }
-                    lineBuffer.removeAll(keepingCapacity: true)
-                }
-            } else {
-                lineBuffer.append(byte)
-                if lineBuffer.count > limits.maxEventBytes {
-                    throw SSEStreamError.eventTooLarge(lineBuffer.count)
-                }
-            }
-        }
-
-        // Flush any final line without a trailing newline.
-        if !lineBuffer.isEmpty,
-           let line = String(data: lineBuffer, encoding: .utf8) {
-            try handleLine(line)
-        }
-
-        // Drain any bytes still held back inside the fallback parser. A stream
-        // that ends without a trailing done-chunk (network cut, malformed
-        // last line) would otherwise swallow the final held-back suffix.
-        if var parser = contentParser {
-            for event in parser.finalize() {
-                _ = try emit(event)
-            }
-            contentParser = parser
-        }
-
-        // Safety net: if the stream ends while thinking is still "open"
-        // (no done-chunk, no empty-thinking transition), still close it out
-        // so consumers don't hang in a thinking-only state.
-        if thinkingOpen {
-            try noteEventYielded()
-            continuation.yield(.thinkingComplete)
-        }
+        var processor = OllamaStreamProcessor(
+            limits: effectiveSSEStreamLimits,
+            config: config,
+            autoDetectedThinkingMarkers: withStateLock { _autoDetectedThinkingMarkers },
+            continuation: continuation,
+            handleUsage: { [weak self] usage in self?.handleUsage(usage) }
+        )
+        try await processor.parse(bytes: bytes)
     }
 
     // MARK: - HTTP Status Validation
@@ -897,270 +440,10 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
             Log.network.debug("Ollama upstream error body: \(errorBody, privacy: .private)")
             let host = self.baseURL?.host()
             let message = CloudErrorSanitizer.sanitize(
-                Self.extractErrorMessage(from: errorBody),
+                OllamaModelProbe.extractErrorMessage(from: errorBody),
                 host: host
             )
             throw CloudBackendError.serverError(statusCode: statusCode, message: message)
-        }
-    }
-
-    // MARK: - NDJSON Parsing
-
-    /// Decoded shape of a single Ollama NDJSON record.
-    ///
-    /// Ollama's two endpoints carry data in different places:
-    /// - `/api/chat` streams put content in `message.content` and reasoning in
-    ///   `message.thinking`.
-    /// - `/api/generate` (non-chat) uses top-level `response` and top-level
-    ///   `thinking`.
-    /// `parseLine` normalises both shapes; consumers read `content` and
-    /// `thinking` without caring which endpoint produced the line.
-    ///
-    /// `evalCount` / `promptEvalCount` are the exact token counts reported by
-    /// the Ollama server. Per Ollama's documented API, these appear on the
-    /// terminal `"done":true` line — `eval_count` is the number of tokens the
-    /// model produced this turn and `prompt_eval_count` is the number of tokens
-    /// in the prompt. Some Ollama-compatible servers also emit a running
-    /// `eval_count` on intermediate lines; parsing it unconditionally lets the
-    /// stream cap visible output precisely when available and falls back to a
-    /// line counter when not.
-    struct ParsedLine {
-        var content: String?
-        var thinking: String?
-        var done: Bool
-        var evalCount: Int?
-        var promptEvalCount: Int?
-        /// Tool calls emitted by the assistant this line, in emission order.
-        /// `nil` when the line carries no `tool_calls` field; an empty array
-        /// is normalised to `nil` so downstream callers can short-circuit on
-        /// `parsed.toolCalls != nil`.
-        var toolCalls: [ToolCall]?
-    }
-
-    /// Parses a single Ollama NDJSON line into a normalised shape.
-    ///
-    /// Returns `nil` for malformed lines so the stream parser can skip them
-    /// the same way it historically skipped unparseable JSON.
-    static func parseLine(_ json: String) -> ParsedLine? {
-        guard let data = json.data(using: .utf8),
-              let parsed = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            return nil
-        }
-
-        let done = (parsed["done"] as? Bool) ?? false
-
-        var content: String?
-        var thinking: String?
-        var toolCalls: [ToolCall]?
-
-        if let message = parsed["message"] as? [String: Any] {
-            // `/api/chat` shape.
-            content = message["content"] as? String
-            thinking = message["thinking"] as? String
-            if let rawCalls = message["tool_calls"] as? [[String: Any]], !rawCalls.isEmpty {
-                toolCalls = rawCalls.compactMap(Self.decodeToolCall)
-                if toolCalls?.isEmpty == true { toolCalls = nil }
-            }
-        }
-
-        // `/api/generate` shape — top-level `response` and `thinking`. If both
-        // `message.content` and top-level `response` are present (shouldn't
-        // happen in practice), chat-shape wins because it arrived first.
-        if content == nil, let response = parsed["response"] as? String {
-            content = response
-        }
-        if thinking == nil, let topThinking = parsed["thinking"] as? String {
-            thinking = topThinking
-        }
-
-        // Usage fields — `eval_count` (output tokens) and `prompt_eval_count`
-        // (prompt tokens). Documented as done-line fields but we parse them
-        // unconditionally so a running-count-emitting server is handled too.
-        let evalCount = parsed["eval_count"] as? Int
-        let promptEvalCount = parsed["prompt_eval_count"] as? Int
-
-        return ParsedLine(
-            content: content,
-            thinking: thinking,
-            done: done,
-            evalCount: evalCount,
-            promptEvalCount: promptEvalCount,
-            toolCalls: toolCalls
-        )
-    }
-
-    // MARK: - Tool-Call Encoding / Decoding Helpers
-
-    /// Serialise a ``ToolDefinition`` into the OpenAI `tools` envelope shape
-    /// Ollama accepts:
-    ///
-    /// ```json
-    /// { "type": "function",
-    ///   "function": { "name": "...", "description": "...", "parameters": {...} } }
-    /// ```
-    ///
-    /// `parameters` round-trips through a JSON encode/decode so the
-    /// `JSONSchemaValue` tree emerges as a plain dictionary/array graph —
-    /// `JSONSerialization` accepts only Foundation primitives and chokes on
-    /// the enum otherwise.
-    static func encodeToolDefinition(_ tool: ToolDefinition) -> [String: Any] {
-        var function: [String: Any] = [
-            "name": tool.name,
-            "description": tool.description,
-        ]
-        if let parameters = Self.foundationJSON(from: tool.parameters) {
-            function["parameters"] = parameters
-        } else {
-            function["parameters"] = ["type": "object", "properties": [String: Any]()]
-        }
-        return [
-            "type": "function",
-            "function": function,
-        ]
-    }
-
-    /// Serialise a ``ToolAwareHistoryEntry`` into Ollama's message shape.
-    ///
-    /// Assistant entries with `toolCalls` get a `tool_calls` array; tool-role
-    /// entries get `tool_call_id`. Plain turns collapse to the same
-    /// `{role, content}` shape the classic history path produces.
-    static func encodeToolAwareEntry(_ entry: ToolAwareHistoryEntry) -> [String: Any] {
-        var obj: [String: Any] = [
-            "role": entry.role,
-            "content": entry.content,
-        ]
-        if let calls = entry.toolCalls, !calls.isEmpty {
-            obj["tool_calls"] = calls.map(Self.encodeToolCall)
-        }
-        if let callId = entry.toolCallId {
-            obj["tool_call_id"] = callId
-        }
-        return obj
-    }
-
-    /// Serialise a single ``ToolCall`` into the OpenAI streaming-compatible
-    /// shape Ollama uses in `message.tool_calls`.
-    ///
-    /// Ollama's server validator parses `arguments` as a JSON object when the
-    /// tool call is fed back in an assistant history entry, so we
-    /// re-hydrate the stored JSON string into a Foundation dictionary before
-    /// emitting. When parsing fails we fall back to an empty object rather
-    /// than shipping a malformed payload — the server will reject the
-    /// request either way, and a clean empty-args call surfaces a more
-    /// actionable error for the host.
-    static func encodeToolCall(_ call: ToolCall) -> [String: Any] {
-        let argumentsValue: Any = Self.parseArgumentString(call.arguments)
-        return [
-            "id": call.id,
-            "type": "function",
-            "function": [
-                "name": call.toolName,
-                "arguments": argumentsValue,
-            ] as [String: Any],
-        ]
-    }
-
-    /// Parse a `ToolCall.arguments` JSON string into the primitive graph
-    /// Ollama expects inside an assistant `tool_calls[]` entry. Falls back
-    /// to an empty object with a log warning on malformed input rather than
-    /// swallowing the error.
-    static func parseArgumentString(_ arguments: String) -> Any {
-        guard let data = arguments.data(using: .utf8) else {
-            Log.inference.warning(
-                "OllamaBackend: tool arguments string was not valid UTF-8 — substituting empty object in history."
-            )
-            return [String: Any]()
-        }
-        do {
-            return try JSONSerialization.jsonObject(with: data, options: [.fragmentsAllowed])
-        } catch {
-            Log.inference.warning(
-                "OllamaBackend: tool arguments string was not valid JSON — substituting empty object in history. error=\(error.localizedDescription, privacy: .public)"
-            )
-            return [String: Any]()
-        }
-    }
-
-    /// Decode one `tool_calls[]` entry from a parsed NDJSON line.
-    ///
-    /// Ollama's streaming format follows the OpenAI shape:
-    /// `{id, type: "function", function: {name, arguments}}`. The `arguments`
-    /// field is sometimes a JSON string (the documented wire shape) and
-    /// sometimes a pre-parsed dictionary (observed on some Ollama builds);
-    /// the decoder handles both and always produces a ``ToolCall`` whose
-    /// `arguments` property is a valid JSON string.
-    ///
-    /// `id` is optional on the wire — some Ollama builds omit it for the
-    /// first tool call in a turn. Synthesise a deterministic fallback from
-    /// the tool name plus a counter suffix when absent so downstream
-    /// call/result pairing still works.
-    static func decodeToolCall(_ raw: [String: Any]) -> ToolCall? {
-        // Two observed shapes on the wire:
-        //   A) {id, type: "function", function: {name, arguments}}  — documented
-        //   B) {id, name, arguments}                                 — some 0.3.x builds
-        // Prefer the nested `function` envelope; fall back to the flat shape
-        // when it's absent so lenient Ollama forks still produce tool events.
-        let nameSource: [String: Any]
-        if let function = raw["function"] as? [String: Any] {
-            nameSource = function
-        } else {
-            nameSource = raw
-        }
-        guard let name = nameSource["name"] as? String, !name.isEmpty else {
-            return nil
-        }
-
-        let id: String
-        if let wireId = raw["id"] as? String, !wireId.isEmpty {
-            id = wireId
-        } else {
-            // Deterministic fallback: ids are only used for id→result pairing
-            // inside one turn, so a name-based placeholder is sufficient.
-            id = "ollama-\(name)-\(UUID().uuidString.prefix(8))"
-        }
-
-        let argumentsString: String
-        if let raw = nameSource["arguments"] as? String {
-            argumentsString = raw
-        } else if let dict = nameSource["arguments"] as? [String: Any] {
-            argumentsString = Self.serialiseArgumentDictionary(dict)
-        } else {
-            argumentsString = "{}"
-        }
-
-        return ToolCall(id: id, toolName: name, arguments: argumentsString)
-    }
-
-    /// Encode a ``JSONSchemaValue`` into the primitive graph
-    /// `JSONSerialization` accepts. Returns `nil` if encoding fails — callers
-    /// are expected to substitute a conservative default.
-    ///
-    /// Delegates to `encodeJSONSchemaToFoundation(_:)` in `ManifoldInference`
-    /// so all backends share one implementation.
-    static func foundationJSON(from value: JSONSchemaValue) -> Any? {
-        encodeJSONSchemaToFoundation(value)
-    }
-
-    /// Serialise an already-parsed arguments dictionary to a JSON string,
-    /// normalising Ollama builds that emit structured `arguments` instead of
-    /// the documented stringified form. Falls back to `"{}"` when
-    /// serialisation fails so ``ToolCall/arguments`` always contains valid
-    /// JSON the registry can decode.
-    static func serialiseArgumentDictionary(_ dict: [String: Any]) -> String {
-        do {
-            let data = try JSONSerialization.data(withJSONObject: dict)
-            if let text = String(data: data, encoding: .utf8) {
-                return text
-            }
-            Log.inference.warning(
-                "OllamaBackend: tool arguments dictionary serialised to non-UTF8 bytes — substituting empty object."
-            )
-            return "{}"
-        } catch {
-            Log.inference.warning(
-                "OllamaBackend: failed to serialise parsed tool arguments — substituting empty object. error=\(error.localizedDescription, privacy: .public)"
-            )
-            return "{}"
         }
     }
 
@@ -1168,125 +451,6 @@ public final class OllamaBackend: SSECloudBackend, CloudBackendURLModelConfigura
 
     public func setToolAwareHistory(_ messages: [ToolAwareHistoryEntry]) {
         withStateLock { self.toolAwareHistory = messages }
-    }
-
-    /// Extracts the assistant content token from an Ollama NDJSON line.
-    ///
-    /// Ollama streaming format (one JSON object per line, no `data:` prefix):
-    /// ```json
-    /// {"model":"llama3","message":{"role":"assistant","content":"Hello"},"done":false}
-    /// ```
-    /// Final chunk has `"done":true` and empty or absent content — we skip it.
-    ///
-    /// This method only surfaces visible content; reasoning-model `thinking`
-    /// fields are handled inline by ``parseResponseStream(bytes:config:continuation:)``
-    /// so they can be emitted as ``GenerationEvent/thinkingToken(_:)`` with
-    /// proper ``GenerationEvent/thinkingComplete`` bracketing. Kept for the
-    /// ``SSEPayloadHandler`` protocol conformance and external callers.
-    static func extractToken(from json: String) -> String? {
-        guard let parsed = parseLine(json) else { return nil }
-        // Skip the final "done" chunk.
-        if parsed.done { return nil }
-        guard let content = parsed.content, !content.isEmpty else { return nil }
-        return content
-    }
-
-    /// Extracts reasoning content from an Ollama NDJSON line, if any.
-    ///
-    /// Returns `nil` when the line carries no `thinking` field or an empty
-    /// one. Exposed for symmetry with ``extractToken(from:)``; streaming
-    /// callers use the inline logic in
-    /// ``parseResponseStream(bytes:config:continuation:)`` to bracket
-    /// thinking emissions with ``GenerationEvent/thinkingComplete``.
-    static func extractThinking(from json: String) -> String? {
-        guard let parsed = parseLine(json),
-              let thinking = parsed.thinking,
-              !thinking.isEmpty else {
-            return nil
-        }
-        return thinking
-    }
-
-    /// Extracts an error message from an Ollama error response body.
-    ///
-    /// Ollama error format: `{"error":"model not found"}`
-    private static func extractErrorMessage(from body: String) -> String? {
-        guard let data = body.data(using: .utf8),
-              let parsed = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-              let message = parsed["error"] as? String else {
-            return nil
-        }
-        return message
-    }
-
-    // MARK: - SSE Payload Handler
-
-    /// Ollama-specific ``SSEPayloadHandler`` for use with ``SSEStreamParser``.
-    ///
-    /// Ollama uses **NDJSON**, not strict SSE — every line is a complete
-    /// JSON object with no `data:` prefix and no blank-line event boundary.
-    /// The shared SSE byte parser cannot strip those frames, so
-    /// ``OllamaBackend`` overrides ``parseResponseStream(bytes:config:continuation:)``
-    /// to walk the byte stream line-by-line instead of via
-    /// ``SSEStreamParser/parse(bytes:limits:eventIDTracker:)``.
-    ///
-    /// What this handler *does* provide is the per-line classification: given
-    /// one NDJSON payload, it surfaces the visible-text and reasoning events
-    /// it carries. Tests for the migration assert against
-    /// ``extractEvents(from:)`` directly, so the handler is the canonical
-    /// per-payload contract even when the production backend bypasses
-    /// ``SSEStreamParser/streamEvents(from:using:limits:)`` for the byte
-    /// loop.
-    struct OllamaPayloadHandler: SSEPayloadHandler {
-        init() {}
-
-        func extractToken(from payload: String) -> String? {
-            OllamaBackend.extractToken(from: payload)
-        }
-
-        /// Maps a single Ollama NDJSON line to the visible-text /
-        /// reasoning-text events it carries.
-        ///
-        /// Emission order matches the on-the-wire field order: any
-        /// `message.thinking` (or top-level `thinking`) on the line emits
-        /// `.thinkingToken` first, then any `message.content` (or top-level
-        /// `response`) emits `.token`.
-        ///
-        /// Tool calls, `done`-line bookkeeping, the per-stream visible /
-        /// thinking caps, and the inline `<think>` fallback all live in the
-        /// stateful byte loop because they require cross-payload state a
-        /// handler cannot model.
-        func extractEvents(from payload: String) -> [GenerationEvent] {
-            guard let parsed = OllamaBackend.parseLine(payload) else { return [] }
-            var events: [GenerationEvent] = []
-            if let thinking = parsed.thinking, !thinking.isEmpty {
-                events.append(.thinkingToken(thinking))
-            }
-            if let content = parsed.content, !content.isEmpty {
-                events.append(.token(content))
-            }
-            return events
-        }
-
-        /// Extracts Ollama's per-turn usage from a single NDJSON payload.
-        ///
-        /// Ollama's documented API places `eval_count` (completion tokens) and
-        /// `prompt_eval_count` (prompt tokens) on the terminal `"done":true`
-        /// line. Returns `nil` when neither field is present so partial/running
-        /// lines don't pollute a consumer that expects "final usage only".
-        func extractUsage(from payload: String) -> (promptTokens: Int?, completionTokens: Int?)? {
-            guard let parsed = OllamaBackend.parseLine(payload) else { return nil }
-            guard parsed.evalCount != nil || parsed.promptEvalCount != nil else {
-                return nil
-            }
-            return (
-                promptTokens: parsed.promptEvalCount,
-                completionTokens: parsed.evalCount
-            )
-        }
-
-        func isStreamEnd(_ payload: String) -> Bool { false }
-        func extractStreamError(from payload: String) -> Error? { nil }
     }
 
     // MARK: - Unload
